refactor: create account service
