+++ src/shared/services/account/protocols/add-account.ts
@@ -1,0 +1,5 @@
+import { AddAccountModel } from '@/modules/account/domain/models/add-account-model';
+
+export interface AddAccountService {
+  addAccount(account: AddAccountModel.Params): Promise<void>;
+}
